+++ packages/atoms/button/index.js
@@ -1,0 +1,9 @@
+import Button from "./Button.svelte";
+
+import Group from "./Group.svelte";
+import GroupItem from "./GroupItem.js";
+import Label from "./Label.svelte";
+
+export default Button;
+
+export { Group, GroupItem, Label };