+++ packages/atoms/date-picker/index.js
@@ -1,0 +1,2 @@
+import DatePicker from "./DatePicker.svelte";
+export default DatePicker;