+++ packages/atoms/logo/Logo.stories.js
@@ -1,31 +1,8 @@
-// import { withKnobs, select, text } from "@storybook/addon-knobs";
-// import { variants } from "@dusk-network/helpers";
-// import DefaultView from "./storybook-views/Default.svelte";
-// import readme from "./readme.stories.md";
-
-// export default {
-//   title: "Components/Atoms/Logo",
-//   parameters: {
-//     notes: readme,
-//   },
-//   decorators: [withKnobs],
-// };
-
-// export const Default = () => ({
-//   Component: DefaultView,
-//   props: {
-//     variant: select("Variant", variants.ATOM.LOGO, variants.ATOM.LOGO.DARK),
-//     href: text("Link", "http://javascript:void(0)"),
-//   },
-// });
-
 import Default from "./storybook-views/Default.svelte";
-
-import "@dusk-network/styles/tailwind.css";
-import Logo from "./Logo.svelte";
+import Logo from "@dusk-network/logo";
 
 export default {
-  title: `components/atoms/Logo`,
+  title: `Components/Atoms/Logo`,
   component: Logo,
   parameters: {
     layout: "centered",