+++ packages/molecules/carousel/index.js
@@ -1,0 +1,4 @@
+import Carousel from "./Carousel.svelte";
+import Item from "./Item.svelte";
+export default Carousel;
+export { Item };