+++ packages/molecules/menu/Menu.stories.js
@@ -1,31 +1,8 @@
-// import { withKnobs, select } from "@storybook/addon-knobs";
-// import { variants, orientations } from "@dusk-network/helpers";
-// import DefaultView from "./storybook-views/Default.svelte";
-// import readme from "./readme.stories.md";
-
-// export default {
-//   title: "Components/Molecules/Menu",
-//   parameters: {
-//     notes: readme,
-//   },
-//   decorators: [withKnobs],
-// };
-
-// export const Default = () => ({
-//   Component: DefaultView,
-//   props: {
-//     orientation: select("Orientation", orientations.MENU, orientations.MENU.VERTICAL),
-//     variant: select("Variant", variants.MOLECULE.MENU, variants.MOLECULE.MENU.BRAND),
-//   },
-// });
-
 import Default from "./storybook-views/Default.svelte";
-
-import "@dusk-network/styles/tailwind.css";
-import Menu from "./Menu.svelte";
+import Menu from "@dusk-network/menu";
 
 export default {
-  title: `components/molecules/Menu`,
+  title: `Components/Molecules/Menu`,
   component: Menu,
   parameters: {
     layout: "centered",