+++ packages/organisms/cookie-banner/index.js
@@ -1,0 +1,2 @@
+import CookieBanner from "./CookieBanner.svelte";
+export default CookieBanner;