+++ packages/organisms/cta/index.js
@@ -1,0 +1,2 @@
+import Cta from "./Cta.svelte";
+export default Cta;