+++ packages/organisms/footer/index.js
@@ -1,0 +1,9 @@
+import Footer from "./Footer.svelte";
+import Maximal from "./Maximal.svelte";
+import Minimal from "./Minimal.svelte";
+import CategoryMenu from "./menu/CategoryMenu.svelte";
+import LegalMenu from "./menu/LegalMenu.svelte";
+import SocialMediaMenu from "./menu/SocialMediaMenu.svelte";
+import Newsletter from "./menu/Newsletter.svelte";
+export default Footer;
+export { Maximal, Minimal, CategoryMenu, LegalMenu, SocialMediaMenu, Newsletter };