+++ packages/organisms/form/index.js
@@ -1,0 +1,2 @@
+import Form from "./Form.svelte";
+export default Form;