+++ packages/organisms/navbar/index.js
@@ -1,0 +1,5 @@
+import Navbar from "./Navbar.svelte";
+import AppsMenu from "./menu/AppsMenu.svelte";
+import LinksMenu from "./menu/LinksMenu.svelte";
+export default Navbar;
+export { AppsMenu, LinksMenu };